+++ src/features/usage-consolidation-accordion/summary-table/column-properties.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { FormattedNumber } from 'react-intl';
 
+import UsageColumnHeader from './usage-column-header';
 import { combineMCLProps } from '../../../components/utilities';
 import {
   formatValue,
@@ -20,7 +21,7 @@ const DEFAULT_SUMMARY_TABLE_COLUMN_WIDTH = {
 };
 
 const DEFAULT_SUMMARY_TABLE_COLUMN_MAPPING = {
-  [DEFAULT_SUMMARY_TABLE_COLUMNS.USAGE]: 'ui-eholdings.usageConsolidation.summary.totalUsage',
+  [DEFAULT_SUMMARY_TABLE_COLUMNS.USAGE]: UsageColumnHeader,
   [DEFAULT_SUMMARY_TABLE_COLUMNS.COST_PER_USE]: 'ui-eholdings.usageConsolidation.summary.costPerUse',
 };
 
@@ -32,20 +33,28 @@ const DEFAULT_FORMATTER = (currency) => ({
 
 export const getCostPerUseFormatter = DEFAULT_FORMATTER;
 
-export const getSummaryTableColumnProperties = (intl, customProps = {}, currency) => {
+export const getSummaryTableColumnProperties = (intl, customProps = {}, args) => {
   const defaultProps = {
     visibleColumns: [...Object.values(DEFAULT_SUMMARY_TABLE_COLUMNS)],
     columnMapping: { ...DEFAULT_SUMMARY_TABLE_COLUMN_MAPPING },
     columnWidths: { ...DEFAULT_SUMMARY_TABLE_COLUMN_WIDTH },
-    formatter: DEFAULT_FORMATTER(currency),
+    formatter: DEFAULT_FORMATTER(args.currency),
   };
 
   const combinedProps = combineMCLProps(defaultProps)(customProps);
 
   const formattedColumnMappingMessages = Object.keys(combinedProps.columnMapping).reduce((memo, currentKey) => {
-    memo[currentKey] = combinedProps.columnMapping[currentKey]
-      ? intl.formatMessage({ id: combinedProps.columnMapping[currentKey] })
-      : null;
+    const columnMappingValue = combinedProps.columnMapping[currentKey];
+
+    if (!columnMappingValue) {
+      memo[currentKey] = null;
+    } else if (typeof columnMappingValue === 'string') {
+      memo[currentKey] = intl.formatMessage({ id: columnMappingValue });
+    } else if (typeof columnMappingValue === 'function') {
+      memo[currentKey] = columnMappingValue(args);
+    } else {
+      memo[currentKey] = columnMappingValue;
+    }
 
     return memo;
   }, {});