+++ src/features/usage-consolidation-accordion/summary-table/summary-table.js
@@ -18,6 +18,7 @@ const propTypes = {
   entityType: PropTypes.string.isRequired,
   id: PropTypes.string.isRequired,
   isExportDisabled: PropTypes.bool,
+  metricType: PropTypes.string,
   onExportTitles: PropTypes.func,
   onViewTitles: PropTypes.func,
 };
@@ -29,6 +30,7 @@ const SummaryTable = ({
   costPerUseType,
   onViewTitles,
   entityType,
+  metricType,
   onExportTitles,
   isExportDisabled,
   ...rest
@@ -59,7 +61,11 @@ const SummaryTable = ({
       <MultiColumnList
         id={id}
         contentData={contentData}
-        {...getSummaryTableColumnProperties(intl, customPropertiesWithFormatter, currency)}
+        {...getSummaryTableColumnProperties(intl, customPropertiesWithFormatter, {
+          currency,
+          metricType,
+          entityType,
+        })}
         {...rest}
       />
     </KeyValue>