+++ src/features/usage-consolidation-accordion/usage-consolidation-content-package.js
@@ -27,6 +27,7 @@ import style from './usage-consolidation-content.css';
 const propTypes = {
   costPerUseData: costPerUseShape.CostPerUseReduxStateShape.isRequired,
   isExportDisabled: PropTypes.bool.isRequired,
+  metricType: PropTypes.string,
   onLoadMoreTitles: PropTypes.func.isRequired,
   onViewTitles: PropTypes.func.isRequired,
   packageId: PropTypes.string.isRequired,
@@ -44,6 +45,7 @@ const UsageConsolidationContentPackage = ({
   onLoadMoreTitles,
   onViewTitles,
   isExportDisabled,
+  metricType,
 }) => {
   const [{ calloutRef }, onExportTitles] = useFetchExportTitlesFromPackage({
     packageId,
@@ -178,6 +180,7 @@ const UsageConsolidationContentPackage = ({
         onExportTitles={onExportTitles}
         onViewTitles={onViewTitles}
         isExportDisabled={isExportDisabled}
+        metricType={metricType}
       />
       <TitlesTable
         costPerUseData={costPerUseData}